django-backend/requirements.txt
PyMySQL==0.9.3
Django<2.0.0,>=1.11.8
django-rest-framework==0.1.0
djangorestframework==3.9.3
requests_oauthlib==1.3.0
google-api-python-client==1.7.11
requests==2.22.0
requests_toolbelt==0.9.1
